Cleanup of imports and simplify custom layout in test
