+++ tests/pyside/test_pyside_renderer.py
@@ -45,6 +45,7 @@ def test_label(qtbot):
 
 
 def test_register_custom_widget_subclass():
+    # Register the class CustomWidget as element Foo
     @cg.PySideRenderer.register_element("Foo")
     class CustomWidget(QtWidgets.QWidget):
         pass
@@ -53,7 +54,12 @@ def test_register_custom_widget_subclass():
         pass
 
     renderer = cg.PySideRenderer(autoshow=False)
-    # renderer.register_element("Foo", typ=CustomWidget)
+
+    with pytest.warns(UserWarning):
+        # It's also possible to register the element with the following line
+        # However, there is already an element registered for Foo, so
+        # a warning will be issued
+        renderer.register_element("Foo", CustomWidget)
 
     foo = renderer.create_element("Foo")
     assert isinstance(foo, CustomWidget)
@@ -66,108 +72,17 @@ def test_register_custom_widget_subclass():
 
 
 def test_register_custom_layout_subclass():
-    class FlowLayout(QtWidgets.QLayout):
+    class HorizontalLayout(QtWidgets.QBoxLayout):
         def __init__(self, parent=None):
-            super().__init__(parent)
-
-            if parent is not None:
-                self.setContentsMargins(QtWidgets.QMargins(0, 0, 0, 0))
-
-            self._item_list = []
-
-        def __del__(self):
-            item = self.takeAt(0)
-            while item:
-                item = self.takeAt(0)
-
-        def addItem(self, item):
-            self._item_list.append(item)
-
-        def count(self):
-            return len(self._item_list)
-
-        def itemAt(self, index):
-            if 0 <= index < len(self._item_list):
-                return self._item_list[index]
-
-            return None
-
-        def takeAt(self, index):
-            if 0 <= index < len(self._item_list):
-                return self._item_list.pop(index)
-
-            return None
-
-        def expandingDirections(self):
-            return QtCore.Qt.Orientation(0)
-
-        def hasHeightForWidth(self):
-            return True
-
-        def heightForWidth(self, width):
-            height = self._do_layout(QtCore.QRect(0, 0, width, 0), True)
-            return height
-
-        def setGeometry(self, rect):
-            super(FlowLayout, self).setGeometry(rect)
-            self._do_layout(rect, False)
-
-        def sizeHint(self):
-            return self.minimumSize()
-
-        def minimumSize(self):
-            size = QtCore.QSize()
-
-            for item in self._item_list:
-                size = size.expandedTo(item.minimumSize())
-
-            size += QtCore.QSize(
-                2 * self.contentsMargins().top(), 2 * self.contentsMargins().top()
-            )
-            return size
-
-        def _do_layout(self, rect, test_only):
-            x = rect.x()
-            y = rect.y()
-            line_height = 0
-            spacing = self.spacing()
-
-            for item in self._item_list:
-                style = item.widget().style()
-                layout_spacing_x = style.layoutSpacing(
-                    QtWidgets.QSizePolicy.PushButton,
-                    QtWidgets.QSizePolicy.PushButton,
-                    QtCore.Qt.Horizontal,
-                )
-                layout_spacing_y = style.layoutSpacing(
-                    QtWidgets.QSizePolicy.PushButton,
-                    QtWidgets.QSizePolicy.PushButton,
-                    QtCore.Qt.Vertical,
-                )
-                space_x = spacing + layout_spacing_x
-                space_y = spacing + layout_spacing_y
-                next_x = x + item.sizeHint().width() + space_x
-                if next_x - space_x > rect.right() and line_height > 0:
-                    x = rect.x()
-                    y = y + line_height + space_y
-                    next_x = x + item.sizeHint().width() + space_x
-                    line_height = 0
-
-                if not test_only:
-                    item.setGeometry(QtCore.QRect(QtCore.QPoint(x, y), item.sizeHint()))
-
-                x = next_x
-                line_height = max(line_height, item.sizeHint().height())
-
-            return y + line_height - rect.y()
+            super().__init__(QtWidgets.QBoxLayout.Direction.LeftToRight, parent)
 
     renderer = cg.PySideRenderer(autoshow=False)
-    renderer.register_layout("flow", FlowLayout)
+    renderer.register_layout("horizontal", HorizontalLayout)
 
     widget = renderer.create_element("widget")
-    renderer.set_attribute(widget, "layout", {"type": "Flow"})
+    renderer.set_attribute(widget, "layout", {"type": "Horizontal"})
 
-    assert isinstance(widget.layout(), FlowLayout)
+    assert isinstance(widget.layout(), HorizontalLayout)
 
 
 def test_widget_add_remove(qtbot):
